+++ Masters Project/Assets/Scripts/Projectiles/RangeAttack.cs
@@ -176,7 +176,7 @@ public abstract class RangeAttack : Attack
         get { return speed; }
     }
 
-    public int GetDamage()
+    public float GetDamage()
     {
         return damage;
     }
